change int to long it

+++ 3-11.cpp
@@ -1,12 +1,12 @@
 #include <stdio.h>
 int main ()
 {
-	int N;
-	scanf("%d", &N);
+	long int N;
+	scanf("%ld", &N);
 	if(N < 0 || N >= 10000000)
 	{
 		printf("Invalid input");
 		return 0;
 	}
-	printf("%d %d", N%3, N%11);
+	printf("%ld %ld", N%3, N%11);
 }
